Use ExitPolicy to compute reconnection delay
When any of the mini-protocol exits (either with an error or otherwise),
we compute reconnection delay which is used by the outbound governor.
It is expected that multiple mini-protocol exits roughly at the same
time, e.g.  in hot to warm transition `chain-sync`, `block-fetch` and
`tx-submission` will terminate around the same time.  For this reason
the `setConnectTimes` is taking the maximum of the existing connect time
and the new one, as well as the `ReconnectDelay` is an additive monoid.